requirements.txt
# Video Processing
opencv-python>=4.8.0
moviepy>=1.0.3
numpy==1.26.0
mediapipe>=0.10.0
torch>=2.1.0
torchvision>=0.16.0

# Audio Processing
faster_whisper==1.0.1
ffmpeg==1.4
ffmpeg_python==0.2.0
pydub>=0.25.1
webrtcvad-wheels>=2.0.11

# LLM APIs
openai>=1.0.0
groq>=0.3.0
tiktoken>=0.5.0

# YouTube
pytube>=15.0.0
youtube-transcript-api>=0.6.1
yt-dlp>=2023.3.4

# Utils
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4==4.12.3
isodate==0.6.1
langdetect>=1.0.9

# Rate Limiting & Retries
tenacity>=8.2.3
ratelimit>=2.2.1
tqdm>=4.66.1

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

--extra-index-url https://download.pytorch.org/whl/cu121
